kaiax/auction: separate eip712 part
