+++ kaiax/auction/bid.go
@@ -29,47 +29,6 @@ import (
 	"github.com/kaiachain/kaia/rlp"
 )
 
-const (
-	auctionType      = "AuctionTx(bytes32 targetTxHash,uint256 blockNumber,address sender,address to,uint256 nonce,uint256 bid,uint256 callGasLimit,bytes data)"
-	EIP712DomainType = "EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"
-	auctionName      = "KAIA_AUCTION"
-	auctionVersion   = "0.0.1"
-)
-
-var (
-	auctionTypeHash    = crypto.Keccak256Hash([]byte(auctionType))
-	eip712TypeHash     = crypto.Keccak256Hash([]byte(EIP712DomainType))
-	auctionNameHash    = crypto.Keccak256Hash([]byte(auctionName))
-	auctionVersionHash = crypto.Keccak256Hash([]byte(auctionVersion))
-)
-
-// EIP712Encoder interface defines the methods required for EIP-712 encoding
-type EIP712Encoder interface {
-	EncodeType() []byte
-	EncodeData() []byte
-}
-
-type EIP712Domain struct {
-	EIP712DomainTypeHash common.Hash
-	NameHash             common.Hash
-	VersionHash          common.Hash
-	ChainId              *big.Int
-	VerifyingContract    common.Address
-}
-
-func (d EIP712Domain) EncodeType() []byte {
-	return d.EIP712DomainTypeHash.Bytes()
-}
-
-func (d EIP712Domain) EncodeData() []byte {
-	encoded := make([]byte, 0)
-	encoded = append(encoded, d.NameHash.Bytes()...)
-	encoded = append(encoded, d.VersionHash.Bytes()...)
-	encoded = append(encoded, common.LeftPadBytes(d.ChainId.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(d.VerifyingContract.Bytes(), 32)...)
-	return encoded
-}
-
 type BidData struct {
 	TargetTxHash  common.Hash    `json:"targetTxHash"`
 	BlockNumber   uint64         `json:"blockNumber"`
@@ -86,6 +45,11 @@ type BidData struct {
 type Bid struct {
 	BidData
 	hash atomic.Value
+}
+
+func (b *Bid) GetEthSignedMessageHash() []byte {
+	data := b.SearcherSig
+	return crypto.Keccak256(fmt.Appendf(nil, "\x19Ethereum Signed Message:\n%d%s", len(data), data))
 }
 
 func (b *Bid) EncodeRLP(w io.Writer) error {
@@ -115,55 +79,6 @@ func rlpHash(x interface{}) (h common.Hash) {
 	rlp.Encode(hw, x)
 	hw.Sum(h[:0])
 	return h
-}
-
-func (b *Bid) EncodeType() []byte {
-	return auctionTypeHash.Bytes()
-}
-
-func (b *Bid) EncodeData() []byte {
-	encoded := make([]byte, 0)
-	encoded = append(encoded, b.TargetTxHash.Bytes()...)
-	encoded = append(encoded, common.LeftPadBytes(common.Int64ToByteBigEndian(b.BlockNumber), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(b.Sender.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(b.To.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(common.Int64ToByteBigEndian(b.Nonce), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(b.Bid.Bytes(), 32)...)
-	encoded = append(encoded, common.LeftPadBytes(common.Int64ToByteBigEndian(b.CallGasLimit), 32)...)
-	encoded = append(encoded, crypto.Keccak256Hash(b.Data).Bytes()...)
-	return encoded
-}
-
-// encodeEIP712 encodes any EIP712Encoder according to EIP-712 specification
-func encodeEIP712(encoder EIP712Encoder) []byte {
-	encoded := make([]byte, 0)
-	encoded = append(encoded, encoder.EncodeType()...)
-	encoded = append(encoded, encoder.EncodeData()...)
-	return crypto.Keccak256Hash(encoded).Bytes()
-}
-
-func (b *Bid) GetHashTypedData(chainId *big.Int, verifyingContract common.Address) []byte {
-	if chainId == nil {
-		return nil
-	}
-
-	domain := EIP712Domain{
-		EIP712DomainTypeHash: eip712TypeHash,
-		NameHash:             auctionNameHash,
-		VersionHash:          auctionVersionHash,
-		ChainId:              chainId,
-		VerifyingContract:    verifyingContract,
-	}
-
-	domainSeparator := encodeEIP712(domain)
-	structHash := encodeEIP712(b)
-
-	return crypto.Keccak256([]byte{0x19, 0x01}, domainSeparator, structHash)
-}
-
-func (b *Bid) GetEthSignedMessageHash() []byte {
-	data := b.SearcherSig
-	return crypto.Keccak256(fmt.Appendf(nil, "\x19Ethereum Signed Message:\n%d%s", len(data), data))
 }
 
 func (b *Bid) ValidateSearcherSig(chainId *big.Int, verifyingContract common.Address) error {
